// packages/datadog-api-client-v2/apis/exception.d.ts
/**
 * Represents an error caused by an api call i.e. it has attributes for a HTTP status code
 * and the returned body object.
 *
 * Example
 * API returns a ErrorMessageObject whenever HTTP status code is not in [200, 299]
 * => ApiException(404, someErrorMessageObject)
 *
 */
export declare class ApiException<T> extends Error {
    code: number;
    body: T;
    constructor(code: number, body: T);
}